requirements.txt
# 코드 파싱
esprima-python==4.0.1  # JavaScript/JSX 파싱
pyjsparser==2.7.1      # 대체 JS 파서 (필요시)

# 임베딩
sentence-transformers==2.2.2
torch==2.0.1

# 벡터 저장
faiss-cpu==1.7.4       # GPU 사용 시 faiss-gpu 설치
numpy==1.24.3

# 유틸리티
tqdm==4.65.0           # 진행률 표시
python-dotenv==1.0.0   # 환경 변수 관리